--- src/ChosenOne.css
@@ -1,13 +1,0 @@
-.chosen-one {
-  background-color: #f9f9f9;
-}
-
-.movie-pic {
-  height: 600px;
-  width: auto;
-}
-
-.chosenOne-details {
-  background-repeat: no-repeat;
-  background-size: 100%;
-}
