+++ src/components/ChosenOne.css
@@ -1,0 +1,41 @@
+.chosen-one {
+  /* background-color: #f9f9f9;
+  background-size: cover; */
+  display: flex;
+  flex-direction: column;
+  z-index: 2;
+}
+
+.video {
+  content: ' ';
+  display: block;
+  position: absolute;
+  width: 100vw;
+  height: 100vh;
+  top: 0px;
+  left: 0;
+  padding-top: 800px;
+}
+
+.movie-pic {
+  height: 600px;
+  width: auto;
+}
+
+.chosenOne-details {
+  /* background-size: 100%; */ 
+  background-repeat: no-repeat;
+  background-size: 100vw 100vh;
+  position: relative;
+  top: 0;
+  left: 0;
+  background-color: #eee8e830;
+  color: white;
+  padding: 200px;
+}
+
+.chosenOne-details h2 {
+  position: relative;
+  z-index: 5;
+  color: white;
+}
